+++ src/main/kotlin/com/gdsc/boolpyeon/user/repository/UserRepository.kt
@@ -3,5 +3,5 @@ package com.gdsc.boolpyeon.user.repository
 import com.gdsc.boolpyeon.user.domain.User
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface UserRepository : JpaRepository<User, Long> {
+interface UserRepository : JpaRepository<User, Int> {
 }
